前端代码提交

+++ EIIM-WEB/src/views/admin/Main.vue
@@ -1,74 +1,73 @@
 <template>
-  <div style="height:100%">
-    <el-container style="height:100%" id="main-container">
-      <el-header id="main-header" height="10%">
-        <el-row type="flex" class="row-bg" align="middle">
-          <el-col :span="20">ddddd</el-col>
-          <el-col :span="4"><i class="iconfont icon-tuichu" @click="logout">退出</i></el-col>
+  <el-container id="admin">
+    <el-header id="main-header">
+      <el-row type="flex" class="row-bg" align="middle">
+        <el-col :span="23" class="title">若森通信软件用户管理平台</el-col>
+        <el-col :span="1"><i class="iconfont icon-tuichu quit" @click="logout">退出</i></el-col>
+      </el-row>
+    </el-header>
+
+    <el-container id="main-body">
+      <el-aside width="400px" :class="'org-list'">
+        <el-row>
+          <el-button type="primary" icon="iconfont icon-add" @click="showOrg">新增</el-button>
         </el-row>
-      </el-header>
-      <el-container height="90%" id="main-body">
-        <el-aside width="400px" :class="'org-list'">
-          <el-row>
-            <el-button type="primary" icon="iconfont icon-add" @click="showOrg">新增</el-button>
-          </el-row>
-          <el-table v-if="orgs && orgs.length > 0"  :data="orgs" @show-header="false" @row-click="selectOrg" :row-style="selectedHighlight" :row-class-name="tableRowClassName" >
-            <el-table-column align="left" prop="name">
-            </el-table-column>
+        <el-table v-if="orgs && orgs.length > 0"  :data="orgs" @show-header="false" @row-click="selectOrg" :row-style="selectedHighlight" :row-class-name="tableRowClassName" >
+          <el-table-column align="left" prop="name" min-width="300px">
+          </el-table-column>
 
-            <el-table-column>
-              <template slot-scope="scope">
-                <i class="iconfont u-btn icon-bianji" @click="handleOrgEdit(scope.$index, scope.row)"></i>
+          <el-table-column>
+            <template slot-scope="scope">
+              <i class="iconfont u-btn icon-bianji" @click="handleOrgEdit(scope.$index, scope.row)"></i>
 
-                <i class="iconfont u-btn icon-shangchu" @click="handleOrgDelete(scope.$index, scope.row)"></i>
-              </template>
-            </el-table-column>
-          </el-table>
-        </el-aside>
+              <i class="iconfont u-btn icon-shangchu" @click="handleOrgDelete(scope.$index, scope.row)"></i>
+            </template>
+          </el-table-column>
+        </el-table>
+      </el-aside>
 
-        <el-main style="background:#fff">
-          <el-container>
-            <el-header style="text-align: right; font-size: 12px">
-              <label id="orgName"></label>
-              <el-button type="primary" icon="el-icon-plus" circle  @click="showUser">新增员工</el-button>
-              <el-button type="danger" icon="iconfont icon-shangchu" circle>删除员工</el-button>
-            </el-header>
-            <el-main>
-              <el-table v-if="org" :data="org.employees" tooltip-effect="dark" @row-dblclick="handleUserEdit">
-                <el-table-column type="selection" mini-width="5%"></el-table-column>
-                <el-table-column prop="fullName" label="姓名" mini-width="15%"></el-table-column>
-                <el-table-column prop="title" label="职务" mini-width="20%"></el-table-column>
-                <el-table-column prop="mobile" label="手机" mini-width="15%"></el-table-column>
-                <el-table-column prop="tel" label="固定电话" mini-width="15%"></el-table-column>
-                <el-table-column prop="mail" label="邮件地址" mini-width="20%"></el-table-column>
-                <el-table-column label="操作" mini-width="10%">
-                  <template slot-scope="scope">
-                    <el-button
-                      size="mini" v-if="!scope.row.account"
-                      @click="showAccount(scope.$index, scope.row)">帐号</el-button>
-                    <el-button
-                      icon="iconfont icon-ban"
-                      size="mini"
-                      type="danger" v-else-if="scope.row.account.state"
-                      @click="disableAccount(scope.$index, scope.row)">禁用</el-button>
-                    <el-button
-                      icon="iconfont icon-qiyong"
-                      size="mini"
-                      type="danger" v-else
-                      @click="enableAccount(scope.$index, scope.row)">启用</el-button>
-                  </template>
-                </el-table-column>
-              </el-table>
-            </el-main>
-          </el-container>
-        </el-main>
-      </el-container>
+      <el-main style="background:#fff">
+        <el-container>
+          <el-header style="text-align: right; font-size: 12px">
+            <label id="orgName"></label>
+            <el-button type="primary" icon="el-icon-plus" circle  @click="showUser">新增用户</el-button>
+            <el-button type="danger" icon="iconfont icon-shangchu" circle>删除用户</el-button>
+          </el-header>
+          <el-main>
+            <el-table v-if="org" :data="org.employees" tooltip-effect="dark" @row-dblclick="handleUserEdit">
+              <el-table-column type="selection" mini-width="5%"></el-table-column>
+              <el-table-column prop="fullName" label="姓名" mini-width="15%"></el-table-column>
+              <el-table-column prop="title" label="职务" mini-width="20%"></el-table-column>
+              <el-table-column prop="mobile" label="手机" mini-width="15%"></el-table-column>
+              <el-table-column prop="phone" label="固定电话" mini-width="10%"></el-table-column>
+              <el-table-column prop="email" label="邮件地址" mini-width="35%"></el-table-column>
+              <el-table-column label="操作" mini-width="5%">
+                <template slot-scope="scope">
+                  <el-button
+                    size="mini" v-if="!scope.row.account"
+                    @click="showAccount(scope.$index, scope.row)">开通帐号</el-button>
+                  <el-button
+                    icon="iconfont icon-ban"
+                    size="mini"
+                    type="danger" v-else-if="scope.row.account.state"
+                    @click="disableAccount(scope.$index, scope.row)">禁用</el-button>
+                  <el-button
+                    icon="iconfont icon-qiyong"
+                    size="mini"
+                    type="danger" v-else
+                    @click="enableAccount(scope.$index, scope.row)">启用</el-button>
+                </template>
+              </el-table-column>
+            </el-table>
+          </el-main>
+        </el-container>
+      </el-main>
     </el-container>
 
     <OrgDialog v-on:setOrgDialogVisible="setOrgDialogVisible" :context="context" :show.sync="showOrgDialog"></OrgDialog>
     <UserDialog v-on:setUserDialogVisible="setUserDialogVisible" :context="context" :show.sync="showUserDialog"></UserDialog>
     <AccountDialog v-on:setAccountDialogVisible="setAccountDialogVisible" :context="context" :show.sync="showAccountDialog"></AccountDialog>
-  </div>
+  </el-container>
 </template>
 
 <script>
@@ -125,12 +124,18 @@ export default {
     },
 
     handleOrgEdit: function (index, row) {
-      this.context = {type: 'ORG', action: 'edit', data: {orgId: row.id}}
+      this.context = {type: 'ORG', action: 'edit', data: row}
       this.showOrgDialog = true
     },
 
-    handleDeleteOrg: function (index, row) {
-      this.$store.dispatch('contact/api_delete_org', row.id)
+    handleOrgDelete: function (index, row) {
+      this.$confirm('此操作将永久删除, 是否继续?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => {
+        this.$store.dispatch('contact/api_delete_org', row.id)
+      })
     },
 
     setOrgDialogVisible (visible) {
@@ -194,7 +199,7 @@ export default {
       row.account.state = false
 
       this.$store.dispatch('account/api_account_disable', row).then(res => {
-        this.store.dispatch('contact/api_update_employee', row)
+        this.$store.dispatch('contact/api_update_employee', row)
       })
     },
 
@@ -225,8 +230,33 @@ export default {
 </script>
 
 <style scoped lang="less">
-  #main-body {
-    background:#fff
+  #admin {
+    width: 100%;
+    height: 100%;
+    margin: 10px 0px 0px 0px;
+    padding: 0;
+  }
+
+  #admin #main-header{
+    height:80px!important;
+  }
+
+  #main-header .title {
+    font-weight: bolder;
+    font-size: 40px;
+    color: #8B5A00;
+  }
+
+  #main-header .quit {
+    font-weight: bolder;
+    font-size: 18px;
+    color: #fff;
+    cursor: pointer;
+  }
+
+  #admin  #main-body {
+    height:850px;
+    background:#fff;
   }
 
   #main-body .org-list{
